split events_driven + other updates
