+++ docs/source/base.rst
@@ -3,7 +3,7 @@ ignite.base
 
 Base module for `ignite.engine.engine` and `ignire.engine.state`
 
-.. currentmodule:: ignite.base
+.. currentmodule:: ignite.base.events_driven
 
 .. autosummary::
     :nosignatures:
